If JSON serialization or deserialization fails for any reason, re-initialize.
The Go Standard Library json.Encoder and json.Decoder are designed for
streaming operations and can be long-lived. However, they're designed to
consume from a single stream. This means that the moment the stream is
determined to be invalid, the decoder renders itself unusable.

To circumvent this, we re-initialize the encoder/decoder on either parsing
failure and/or stream-operation failure (read/write errors).